chore(popular movies): renaming components
