+++ src/navigator/index.ts
@@ -1,6 +1,1 @@
 export { StackNavigator } from './StackNavigator';
-
-export type AppStackParamList = {
-  'Popular Movies': {};
-  'Movie Details': {};
-};
