+++ src/screens/PopularMovies/PopularMovies.tsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import { useQuery } from 'react-apollo';
+
+import { Loading, Error } from 'components';
+import { PopularMoviesList } from './components';
+import { POPULAR_MOVIES } from './PopularMovies.graphql';
+import { Movies } from './PopularMovies.types';
+import * as Styled from './PopularMovies.style';
+
+export const PopularMovies: React.FC = () => {
+  const { loading, data, error } = useQuery<Movies>(POPULAR_MOVIES);
+
+  if (loading) return <Loading />;
+
+  if (error) return <Error />;
+
+  const moviesList = data?.popularMovies.map(movie => (
+    <PopularMoviesList movie={movie} key={movie.id} />
+  ));
+
+  return (
+    <Styled.Container>
+      <Styled.Movies>{moviesList}</Styled.Movies>
+    </Styled.Container>
+  );
+};
