--- src/screens/PopularMovies/container/Container.types.ts
@@ -1,8 +1,0 @@
-export type PopularMovies = {
-  popularMovies: {
-    id: number;
-    title: string;
-    posterPath: string;
-    backdropPath: string;
-  }[];
-};
