--- src/screens/PopularMovies/container/index.ts
@@ -1,1 +1,0 @@
-export { PopularMoviesContainer } from './Container';
